==> data_preprocessing/1_news_data_loading.ipynb <==
# Cite from Xinyi6
import os
import json
import pandas as pd


# load all files in all sub-folders under the parent directory
def list_files(dirpath):
    for dirname, dirnames, filenames in os.walk(dirpath):
        for filename in filenames:
            yield os.path.join(dirname, filename)


# Load all .json files use json.loads(), basically each .json file only contains 1 line
json_list = []
dirpath = '../data/us-financial-news-articles/'
for filePath in list_files(dirpath):
    if filePath.endswith('.json'):
        with open(filePath) as f:
            for line in f:
                data = json.loads(line)
                json_list.append([
                    data['published'],
                    data['thread']['site'],
                    data['title'],
                    data['text'],
                    data['url']])


# Make sure the length of json list matches the total files
len(json_list)


# Convert Json to DataFrame in order to perform data analysis
col_names = ['published_date', 'source_name', 'title', 'body', 'url']
df = pd.DataFrame(json_list, columns=col_names)
df.shape


df.head()


# Sort the data by date
df = df.sort_values(by=['published_date'], ascending=True)
df.head()


df = df.reset_index(inplace=False)
del df['index']


df.head()


# Export to .csv file
# df.to_csv('../data/large_data/us_financial_news_articles_2018.csv')
