==> data_preprocessing/4_OECD_data_aggregation.ipynb <==
import pandas as pd
from sklearn.preprocessing import MinMaxScaler


CLI_df = pd.read_csv(
    '../data/oecd_data/CLI.csv')[['INDICATOR', 'TIME', 'Value']]
BCI_df = pd.read_csv(
    '../data/oecd_data/BCI.csv')[['INDICATOR', 'TIME', 'Value']]
CCI_df = pd.read_csv(
    '../data/oecd_data/CCI.csv')[['INDICATOR', 'TIME', 'Value']]


CLI_df


src_df = pd.read_csv('../data/source_price.csv')
src_df


for index, row in src_df.iterrows():
    date = row['date'].rsplit('/', 1)[0]
    src_df.at[index, 'CLI'] = CLI_df[CLI_df['TIME']
                                     == date]['Value'].item()-100
    src_df.at[index, 'BCI'] = BCI_df[BCI_df['TIME']
                                     == date]['Value'].item()-100
    src_df.at[index, 'CCI'] = CCI_df[CCI_df['TIME']
                                     == date]['Value'].item()-100


src_df


src_df.to_csv('../data/source_price_sentiment_oecd.csv', index=False)
